Add text formatting for script event commands
Added some code to allow arbitrary formatting of commands in scripts using the values from an event.
This should be failsafe as the formatting used implements a custom dict object that will return the key if the format cannot find it in the string to format.

Also moved the decode_if_bytes and encode_if_str method to shared as it was needed here too