Add config for lang file path for console tracking
Added a config option to point to a lang file for console tracking. This will revive the PlayerDeath event!

To get a lang file for your version of minecraft open the jar in an archive manager (WinRAR, 7Zip etc.) and look for `assets/minecraft/lang/en_us.json` or it might be a `.lang` file, either should work.