Replace usage of imp with importlib
Replace the usage of imp for importing modules with importlib.

This fixes compatability with Python 3.12+

This code is ripped straight from the python docs for loading a module from the source which is what was used in mark2 originally. This code _could_ cause unforseen problems so I will keep an eye on if this breaks stuff

Closes #172